Integrate: fix int64

+++ core/block/editor/state/state.go
@@ -1762,7 +1762,7 @@ func (s *State) GetChangedStoreKeys(prefixPath ...string) (paths [][]string) {
 func (s *State) Layout() (model.ObjectTypeLayout, bool) {
 	if det := s.Details(); det != nil {
 		if det.Has(bundle.RelationKeyLayout) {
-			return model.ObjectTypeLayout(det.GetIntOrDefault(bundle.RelationKeyLayout, 0)), true
+			return model.ObjectTypeLayout(det.GetInt64OrDefault(bundle.RelationKeyLayout, 0)), true
 		}
 	}
 	return 0, false
